+++ MortalFighter/ValidadorDelColor.h
@@ -2,12 +2,12 @@
 class ValidadorDelColor
 {
 private:
-	std::vector<int> colorAlternativo;
+	std::vector<double> colorAlternativo;
 
 public:
 	ValidadorDelColor();
 	void validarColorDesde(Json::Value color_alternativo);
-	std::vector<int> getColorAlternativo();
+	std::vector<double> getColorAlternativo();
 	~ValidadorDelColor();
 };
